add change to mapper

+++ microservice-client/src/main/java/com/co/technicaltest/neoris/client/models/dto/ClientResponseDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public record ClientResponseDTO(Long id, String name, Gender gender, Integer age,
                                 String identification, String address, String phone,
-                                Boolean status, List<Integer> clientAccounts, List<Account> clientAccountsDetail) {
+                                Boolean status, List<Long> clientAccounts, List<Account> clientAccountsDetail) {
 }
